Add getgrouplist method to Etc
Add a `getgrouplist` method to the Etc module, which behaves similarly
to the glibc method, with some significant differences.

The glibc method returns all GIDs for groups a user is part of. It
also requires a `gid_t` parameter which is usually the group ID from
the user's passwd record. If the group is not among the groups defined
for the user, it's still included in the `getgrouplist()` output.[1]

The Ruby version does the following:
- calls `getpwnam()` to get the gid of the queried user
- calls `getgrouplist()` twice with the user and gid (the first call
is just to get the number of groups)
- calls `getgrgid()` for each group ID returned by `getgrouplist()`
- removes the default group if the queried user is not a member of it
- returns the groups in the form of an array of Etc::Group structs

[1] https://man7.org/linux/man-pages/man3/getgrouplist.3.html

Signed-off-by: Gabriel Nagy <gabriel.nagy@puppet.com>